refactor(levm): move opcodes to their corresponding folder (#863) (#910)
**Motivation**

Implement the issue "Move opcodes to their corresponding folder" #863 

**Description**

Moved the opcodes from block.rs in environment.rs accordings to the list
commented at the beginning of each file.
The modification was verified by running the tests via 'make test'.

Resolves #863

Closes #863
