+++ crates/vm/levm/src/opcode_handlers/block.rs
@@ -1,9 +1,7 @@
 use crate::{
     block::LAST_AVAILABLE_BLOCK_LIMIT,
-    constants::{call_opcode, BLOB_BASE_FEE_UPDATE_FRACTION, MIN_BASE_FEE_PER_BLOB_GAS, WORD_SIZE},
-    vm::word_to_address,
+    constants::{BLOB_BASE_FEE_UPDATE_FRACTION, MIN_BASE_FEE_PER_BLOB_GAS},
 };
-use sha3::{Digest, Keccak256};
 
 // Block Information (11)
 // Opcodes: BLOCKHASH, COINBASE, TIMESTAMP, NUMBER, PREVRANDAO, GASLIMIT, CHAINID, SELFBALANCE, BASEFEE, BLOBHASH, BLOBBASEFEE
@@ -176,244 +174,6 @@ impl VM {
         let blob_base_fee = self.get_blob_gasprice();
 
         current_call_frame.stack.push(blob_base_fee)?;
-
-        Ok(OpcodeSuccess::Continue)
-    }
-
-    // ADDRESS operation
-    pub fn op_address(
-        &mut self,
-        current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
-        self.increase_consumed_gas(current_call_frame, gas_cost::ADDRESS)?;
-
-        let addr = if current_call_frame.delegate.is_some() {
-            current_call_frame.msg_sender
-        } else {
-            current_call_frame.code_address
-        };
-
-        current_call_frame.stack.push(U256::from(addr.as_bytes()))?;
-
-        Ok(OpcodeSuccess::Continue)
-    }
-
-    // BALANCE operation
-    pub fn op_balance(
-        &mut self,
-        current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
-        self.increase_consumed_gas(current_call_frame, gas_cost::BALANCE)?;
-
-        let addr = current_call_frame.stack.pop()?;
-
-        let balance = self.db.balance(&word_to_address(addr));
-        current_call_frame.stack.push(balance)?;
-
-        Ok(OpcodeSuccess::Continue)
-    }
-
-    // ORIGIN operation
-    pub fn op_origin(
-        &mut self,
-        current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
-        self.increase_consumed_gas(current_call_frame, gas_cost::ORIGIN)?;
-
-        let origin = self.env.origin;
-        current_call_frame
-            .stack
-            .push(U256::from(origin.as_bytes()))?;
-
-        Ok(OpcodeSuccess::Continue)
-    }
-
-    // CALLER operation
-    pub fn op_caller(
-        &mut self,
-        current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
-        self.increase_consumed_gas(current_call_frame, gas_cost::CALLER)?;
-
-        let caller = current_call_frame.msg_sender;
-        current_call_frame
-            .stack
-            .push(U256::from(caller.as_bytes()))?;
-
-        Ok(OpcodeSuccess::Continue)
-    }
-
-    // CALLVALUE operation
-    pub fn op_callvalue(
-        &mut self,
-        current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
-        self.increase_consumed_gas(current_call_frame, gas_cost::CALLVALUE)?;
-
-        let callvalue = current_call_frame.msg_value;
-
-        current_call_frame.stack.push(callvalue)?;
-
-        Ok(OpcodeSuccess::Continue)
-    }
-
-    // CODESIZE operation
-    pub fn op_codesize(
-        &mut self,
-        current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
-        self.increase_consumed_gas(current_call_frame, gas_cost::CODESIZE)?;
-
-        current_call_frame
-            .stack
-            .push(U256::from(current_call_frame.bytecode.len()))?;
-
-        Ok(OpcodeSuccess::Continue)
-    }
-
-    // CODECOPY operation
-    pub fn op_codecopy(
-        &mut self,
-        current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
-        let dest_offset: usize = current_call_frame
-            .stack
-            .pop()?
-            .try_into()
-            .map_err(|_| VMError::VeryLargeNumber)?;
-        let offset: usize = current_call_frame
-            .stack
-            .pop()?
-            .try_into()
-            .map_err(|_| VMError::VeryLargeNumber)?;
-        let size: usize = current_call_frame
-            .stack
-            .pop()?
-            .try_into()
-            .map_err(|_| VMError::VeryLargeNumber)?;
-
-        let minimum_word_size = (size + WORD_SIZE - 1) / WORD_SIZE;
-
-        let memory_expansion_cost = current_call_frame
-            .memory
-            .expansion_cost(dest_offset + size)?;
-
-        let gas_cost = gas_cost::CODECOPY_STATIC
-            + gas_cost::CODECOPY_DYNAMIC_BASE * minimum_word_size
-            + memory_expansion_cost;
-
-        self.increase_consumed_gas(current_call_frame, gas_cost)?;
-
-        let code = current_call_frame.bytecode.slice(offset..offset + size);
-
-        current_call_frame.memory.store_bytes(dest_offset, &code);
-
-        Ok(OpcodeSuccess::Continue)
-    }
-
-    // GASPRICE operation
-    pub fn op_gasprice(
-        &mut self,
-        current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
-        self.increase_consumed_gas(current_call_frame, gas_cost::GASPRICE)?;
-
-        current_call_frame.stack.push(self.env.gas_price)?;
-
-        Ok(OpcodeSuccess::Continue)
-    }
-
-    // EXTCODESIZE operation
-    pub fn op_extcodesize(
-        &mut self,
-        current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
-        let address = word_to_address(current_call_frame.stack.pop()?);
-        let gas_cost = if self.accrued_substate.warm_addresses.contains(&address) {
-            call_opcode::WARM_ADDRESS_ACCESS_COST
-        } else {
-            call_opcode::COLD_ADDRESS_ACCESS_COST
-        };
-
-        self.increase_consumed_gas(current_call_frame, gas_cost)?;
-
-        let code_size = self.db.get_account_bytecode(&address).len();
-        current_call_frame.stack.push(code_size.into())?;
-
-        Ok(OpcodeSuccess::Continue)
-    }
-
-    // EXTCODECOPY operation
-    pub fn op_extcodecopy(
-        &mut self,
-        current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
-        let address = word_to_address(current_call_frame.stack.pop()?);
-        let dest_offset: usize = current_call_frame
-            .stack
-            .pop()?
-            .try_into()
-            .map_err(|_| VMError::VeryLargeNumber)?;
-        let offset: usize = current_call_frame
-            .stack
-            .pop()?
-            .try_into()
-            .map_err(|_| VMError::VeryLargeNumber)?;
-        let size: usize = current_call_frame
-            .stack
-            .pop()?
-            .try_into()
-            .map_err(|_| VMError::VeryLargeNumber)?;
-
-        let minimum_word_size = (size + WORD_SIZE - 1) / WORD_SIZE;
-        let memory_expansion_cost = current_call_frame
-            .memory
-            .expansion_cost(dest_offset + size)?;
-        let address_access_cost = if self.accrued_substate.warm_addresses.contains(&address) {
-            call_opcode::WARM_ADDRESS_ACCESS_COST
-        } else {
-            call_opcode::COLD_ADDRESS_ACCESS_COST
-        };
-        let gas_cost = gas_cost::EXTCODECOPY_DYNAMIC_BASE * minimum_word_size
-            + memory_expansion_cost
-            + address_access_cost;
-
-        self.increase_consumed_gas(current_call_frame, gas_cost)?;
-
-        let mut code = self.db.get_account_bytecode(&address);
-        if code.len() < offset + size {
-            let mut extended_code = code.to_vec();
-            extended_code.resize(offset + size, 0);
-            code = Bytes::from(extended_code);
-        }
-        current_call_frame
-            .memory
-            .store_bytes(dest_offset, &code[offset..offset + size]);
-
-        Ok(OpcodeSuccess::Continue)
-    }
-
-    // EXTCODEHASH operation
-    pub fn op_extcodehash(
-        &mut self,
-        current_call_frame: &mut CallFrame,
-    ) -> Result<OpcodeSuccess, VMError> {
-        let address = word_to_address(current_call_frame.stack.pop()?);
-        let gas_cost = if self.accrued_substate.warm_addresses.contains(&address) {
-            call_opcode::WARM_ADDRESS_ACCESS_COST
-        } else {
-            call_opcode::COLD_ADDRESS_ACCESS_COST
-        };
-
-        self.increase_consumed_gas(current_call_frame, gas_cost)?;
-
-        let code = self.db.get_account_bytecode(&address);
-        let mut hasher = Keccak256::new();
-        hasher.update(code);
-        let result = hasher.finalize();
-        current_call_frame
-            .stack
-            .push(U256::from_big_endian(&result))?;
 
         Ok(OpcodeSuccess::Continue)
     }
